+++ internal/framework/fetch/fetch.go
@@ -33,26 +33,6 @@ const (
 	checksumFileSuffix = ".sha256"
 )
 
-// ChecksumMismatchError represents an error when the calculated checksum doesn't match the expected checksum.
-// This type of error should not trigger retries as it indicates data corruption or tampering.
-type ChecksumMismatchError struct {
-	Expected string
-	Actual   string
-}
-
-func (e *ChecksumMismatchError) Error() string {
-	return fmt.Sprintf("checksum mismatch: expected %s, got %s", e.Expected, e.Actual)
-}
-
-// HTTPStatusError represents an HTTP status code error for retry logic.
-type HTTPStatusError struct {
-	StatusCode int
-}
-
-func (e *HTTPStatusError) Error() string {
-	return fmt.Sprintf("unexpected status code: %d", e.StatusCode)
-}
-
 // RetryBackoffType defines supported backoff strategies.
 type RetryBackoffType string
 
@@ -197,16 +177,7 @@ func (f *DefaultFetcher) GetRemoteFile(targetURL string) ([]byte, error) {
 		return nil, fmt.Errorf("failed to fetch HTTP file after retries: %w", err)
 	}
 
-	if result != nil {
-		return result, nil
-	}
-
-	// This case should ideally not be reached, but as a fallback, return the last known error.
-	if lastErr != nil {
-		return nil, lastErr
-	}
-
-	return nil, fmt.Errorf("failed to fetch HTTP file %s: unknown error", targetURL)
+	return result, nil
 }
 
 // getFileContent fetches content via HTTP(S).
